+++ packages/components/tree-select/src/select.ts
@@ -1,39 +1,42 @@
-import { getCurrentInstance, onMounted, toRefs } from 'vue'
+import { computed, toRefs } from 'vue'
 import { pick } from 'lodash-unified'
 import ElSelect from '../../select/src/select.vue'
-import type { ComponentInternalInstance, Ref } from 'vue'
-import type { ComponentPropsType, RecordRef } from './type'
-
-type ElSelectPropsType = RecordRef<ComponentPropsType<typeof ElSelect>>
+import type ElTree from '../../tree/src/tree.vue'
+import type { Ref } from 'vue'
 
 export const useSelect = (
-  allProps,
-  attrs,
-  select: Ref<InstanceType<typeof ElSelect> | undefined>
-) => {
-  const props = pick(
-    toRefs(allProps),
-    Object.keys(ElSelect.props)
-  ) as ElSelectPropsType
-
-  const listeners = pick(attrs, [
-    'onChange',
-    'onVisibleChange',
-    'onRemoveTag',
-    'onClear',
-    'onBlur',
-    'onFocus',
-  ])
-
-  // set same methods when mounted
-  onMounted(() => {
-    const methods = ['focus', 'blur']
-    const instance = getCurrentInstance() as ComponentInternalInstance
-    methods.forEach((item) => (instance[item] = select.value?.[item]))
-  })
-
-  return {
-    props,
-    listeners,
+  props,
+  { attrs },
+  {
+    tree,
+    key,
+  }: {
+    select: Ref<InstanceType<typeof ElSelect> | undefined>
+    tree: Ref<InstanceType<typeof ElTree> | undefined>
+    key: Ref<string>
   }
+) => {
+  const result = {
+    ...pick(toRefs(props), Object.keys(ElSelect.props)),
+    ...attrs,
+    valueKey: key,
+    popperClass: computed(() => {
+      const classes = ['el-tree-select__popper']
+      if (props.popperClass) classes.push(props.popperClass)
+      return classes.join(' ')
+    }),
+    filterMethod: (keyword = '') => {
+      tree.value?.filter(keyword)
+    },
+    // clear filter text when visible change
+    onVisibleChange: (visible: boolean) => {
+      ;(attrs.onVisibleChange as (visible: boolean) => void)?.(visible)
+
+      if (props.filterable && visible) {
+        result.filterMethod()
+      }
+    },
+  }
+
+  return result
 }
